feat(mattress_and_soldout): mattress and sold out complete

+++ make_my_trip/lib/features/home_page/data/models/ToursModel.dart
@@ -25,7 +25,7 @@ class ToursModel {
   ToursModel.fromJson(dynamic json) {
     _id = json['_id'];
     _tourName = json['tour_name'];
-    _rating = json['rating'];
+    _rating = json['rating'].toDouble();
     _price = json['price'];
     if (json['images'] != null) {
       _images = [];
